requirements.txt
openai==1.40.0
anthropic==0.34.0
requests==2.31.0
python-dotenv==1.0.1
googlesearch-python==1.2.3
schedule==1.2.1
pydantic==2.8.2
beautifulsoup4==4.12.3
lxml==5.2.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.122.0

# Video Generation Dependencies
pillow==10.3.0
moviepy==1.0.3

# Option 1: AI Voice (ElevenLabs)
elevenlabs==1.1.0

# Option 2: Avatar (D-ID)
# requests (already included)

# Option 3: Free TTS
gTTS==2.5.1
edge-tts==6.1.10

# New Features Dependencies
# Multiple News Sources
feedparser==6.0.11  # RSS feeds
newsapi-python==0.2.7  # NewsAPI integration

# Subtitle Generation
openai-whisper==20231117  # Optional: for audio transcription
# Note: Whisper requires ffmpeg to be installed separately

# Thumbnail Generation (Pillow already included)
# For AI thumbnails, OpenAI is already included

